display: Move bulk of UI out of cc-display-panel
This patch moves all the UI code except that related to the "apply"
button and current monitor detection out into a seperate panel.
The portion moved out is common to other components which wish to
inteact with the displays and their layout.

This code does not compile as-is, there is still some cleanup work
to be done.